Everything sample: play music from start
Playing the music right from the start tests the new sound on WASM. In
the browser sound only starts after the first user interaction (mouse
or keyboard presses). This means that the sample code calls PlaySound
right at the start but the music does not start. After the first user
interaction the music will set in at the right offset as if it were
played silently from the beginning.